Break up and reorganize files
