+++ db/queries/instantiations/data_sources/put.py
@@ -10,9 +10,7 @@ from db.queries.builder.mixins.pending_event.data_source import (
 )
 from db.queries.builder.mixins.record_type import RecordTypeMixin
 from middleware.enums import PermissionsEnum
-from middleware.schema_and_dto.schemas.data_sources.base import (
-    EntryCreateUpdateRequestDTO,
-)
+from middleware.schema_and_dto.dtos.entry_create_update_request import EntryCreateUpdateRequestDTO
 from middleware.util.type_conversion import dict_enums_to_values
