+++ endpoints/instantiations/data_sources_/_shared/schemas/base.py
@@ -1,5 +1,4 @@
 from marshmallow import Schema, fields
-from pydantic import BaseModel
 
 from db.enums import (
     AgencyAggregation,
@@ -10,10 +9,7 @@ from db.enums import (
     URLStatus,
     ApprovalStatus,
 )
-from middleware.schema_and_dto.non_dto_dataclasses import DTOPopulateParameters
-from middleware.schema_and_dto.schemas.common.custom_fields import DataField
 from middleware.schema_and_dto.util import get_json_metadata
-from utilities.enums import SourceMappingEnum
 
 
 class DataSourceBaseSchema(Schema):
@@ -267,27 +263,3 @@ class DataSourceBaseSchema(Schema):
     )
 
 
-class EntryDataRequestSchema(Schema):
-    entry_data = DataField(
-        required=True,
-        metadata={
-            "source": SourceMappingEnum.JSON,
-            "description": "The entry data field for adding and updating entries",
-        },
-    )
-
-
-class EntryCreateUpdateRequestDTO(BaseModel):
-    """
-    Contains data for creating or updating an entry
-    """
-
-    entry_data: dict
-
-    @classmethod
-    def get_dto_populate_parameters(cls) -> DTOPopulateParameters:
-        return DTOPopulateParameters(
-            dto_class=EntryCreateUpdateRequestDTO,
-            source=SourceMappingEnum.JSON,
-            validation_schema=EntryDataRequestSchema,
-        )
