+++ endpoints/schema_config/instantiations/data_sources/by_id/put.py
@@ -1,5 +1,5 @@
 from endpoints.schema_config.helpers import get_put_resource_endpoint_schema_config
-from middleware.schema_and_dto.schemas.data_sources.base import EntryDataRequestSchema
+from middleware.schema_and_dto.schemas.data_sources.entry_data_request import EntryDataRequestSchema
 from middleware.schema_and_dto.schemas.data_sources.put import DataSourcesPutSchema
 
 DataSourcesByIDPutEndpointSchemaConfig = get_put_resource_endpoint_schema_config(
