+++ tests/integration/notifications/event_to_pending/data_sources/put/manager.py
@@ -1,9 +1,7 @@
 from db.client.core import DatabaseClient
 from db.enums import ApprovalStatus
 from middleware.enums import PermissionsEnum
-from middleware.schema_and_dto.schemas.data_sources.base import (
-    EntryCreateUpdateRequestDTO,
-)
+from middleware.schema_and_dto.dtos.entry_create_update_request import EntryCreateUpdateRequestDTO
 from tests.integration.notifications.event_to_pending.data_sources.manager import (
     EventToPendingDataSourcesTestManager,
 )
