+++ tests/middleware/test_common_schemas_and_dtos.py
@@ -1,9 +1,7 @@
 import pytest
 from marshmallow import ValidationError
 
-from middleware.schema_and_dto.schemas.data_sources.base import (
-    EntryDataRequestSchema,
-)
+from middleware.schema_and_dto.schemas.data_sources.entry_data_request import EntryDataRequestSchema
 
 
 def test_entry_data_request_schema_success():
